refactor(tasks): extract TaskSorter from TaskRegistry
Move the sort comparators and section-preserving bucketing out of the
476-LOC TaskRegistry god-class into a pure, Sendable `TaskSorter` value
under Tasks/Registry/. TaskRegistry now holds an injected sorter and
delegates ordering in `tasks`/`completedTasks`; behavior is unchanged.

Split the sort tests: pure ordering mechanics move to TaskSorterTests
(fast, no registry/provider/async), and the query-scope -> section
mapping moves to TaskQueryScopeTests. TaskRegistrySortTests is removed.

First slice of the #404 registry split (design doc 0005, D1 / Finding M).
