+++ app/Taskmato/Tasks/TaskRegistry.swift
@@ -50,12 +50,16 @@ final class TaskRegistry {
   }
 
   private let defaults: UserDefaults
+  private let sorter: TaskSorter
   private static let enabledKey = "taskRegistry.enabledProviderIDs"
   private static let selectionKey = "taskRegistry.selection"
 
-  /// - Parameter defaults: `UserDefaults` store for persistence. Override in tests.
-  init(defaults: UserDefaults = .standard) {
+  /// - Parameters:
+  ///   - defaults: `UserDefaults` store for persistence. Override in tests.
+  ///   - sorter: The task sorter used to order query results.
+  init(defaults: UserDefaults = .standard, sorter: TaskSorter = TaskSorter()) {
     self.defaults = defaults
+    self.sorter = sorter
 
     // One-shot migration: remove abandoned list-scope blobs from prior versions.
     defaults.removeObject(forKey: "taskRegistry.providerListScopes")
@@ -215,7 +219,7 @@ final class TaskRegistry {
         return (tasks: [], errors: [])
       }
       let items = (try? await provider.tasks(in: list)) ?? []
-      return (tasks: sortedByField(items, field: field, direction: direction), errors: [])
+      return (tasks: sorter.sorted(items, by: field, direction: direction), errors: [])
 
     case .crossProvider(let filter):
       let (all, errors) = await globalFanOut(providers: active)
@@ -232,7 +236,7 @@ final class TaskRegistry {
         filtered = all.filter { $0.title.localizedCaseInsensitiveContains(titleQuery) }
       }
       return (
-        tasks: sortedByField(filtered, field: field, direction: direction, preserveSections: false),
+        tasks: sorter.sorted(filtered, by: field, direction: direction, preserveSections: false),
         errors: errors
       )
     }
@@ -294,7 +298,7 @@ final class TaskRegistry {
     }
 
     return (
-      tasks: sortedByField(filtered, field: field, direction: direction, preserveSections: false),
+      tasks: sorter.sorted(filtered, by: field, direction: direction, preserveSections: false),
       errors: providerErrors
     )
   }
@@ -396,88 +400,6 @@ final class TaskRegistry {
     }
 
     return (merged, providerErrors)
-  }
-
-  /// Sorts `items` by `field`/`direction`, applying the sort within each `(list.id, section)`
-  /// group in encounter order. Groups are then flattened back to a single array, preserving
-  /// the section order the providers produced.
-  private func sortedByField(
-    _ items: [TaskItem],
-    field: TaskSortField,
-    direction: TaskSortDirection,
-    preserveSections: Bool = true
-  ) -> [TaskItem] {
-    guard preserveSections else {
-      return items.sorted { compareItems($0, $1, field: field, direction: direction) }
-    }
-
-    struct SectionKey: Hashable {
-      let listID: String
-      let section: String?
-    }
-
-    var ordered: [SectionKey] = []
-    var bySection: [SectionKey: [TaskItem]] = [:]
-
-    for item in items {
-      let key = SectionKey(listID: item.list?.id ?? "", section: item.section)
-      if bySection[key] == nil { ordered.append(key) }
-      bySection[key, default: []].append(item)
-    }
-
-    return ordered.flatMap { key in
-      (bySection[key] ?? []).sorted { compareItems($0, $1, field: field, direction: direction) }
-    }
-  }
-
-  private func compareItems(
-    _ lhs: TaskItem, _ rhs: TaskItem, field: TaskSortField, direction: TaskSortDirection
-  ) -> Bool {
-    let asc = direction == .ascending
-    switch field {
-    case .dueDate:
-      return compareDatesNilLast(lhs.dueDate, rhs.dueDate, ascending: asc)
-        ?? compareTitlesAscending(lhs.title, rhs.title)
-        ?? compareRefs(lhs.id, rhs.id)
-    case .priority:
-      if lhs.priority != rhs.priority {
-        return asc ? lhs.priority < rhs.priority : lhs.priority > rhs.priority
-      }
-      return compareDatesNilLast(lhs.dueDate, rhs.dueDate, ascending: true)
-        ?? compareTitlesAscending(lhs.title, rhs.title)
-        ?? compareRefs(lhs.id, rhs.id)
-    case .title:
-      let cmp = lhs.title.localizedStandardCompare(rhs.title)
-      if cmp != .orderedSame { return asc ? cmp == .orderedAscending : cmp == .orderedDescending }
-      return compareRefs(lhs.id, rhs.id)
-    case .creationDate:
-      return compareDatesNilLast(lhs.createdAt, rhs.createdAt, ascending: asc)
-        ?? compareTitlesAscending(lhs.title, rhs.title)
-        ?? compareRefs(lhs.id, rhs.id)
-    }
-  }
-
-  /// Compares two optional dates with nil-last semantics. Returns `nil` when both are equal or both nil.
-  private func compareDatesNilLast(_ lhs: Date?, _ rhs: Date?, ascending: Bool) -> Bool? {
-    switch (lhs, rhs) {
-    case (let lhsDate?, let rhsDate?):
-      guard lhsDate != rhsDate else { return nil }
-      return ascending ? lhsDate < rhsDate : lhsDate > rhsDate
-    case (.some, .none): return true
-    case (.none, .some): return false
-    case (.none, .none): return nil
-    }
-  }
-
-  /// Ascending title comparison using `localizedStandardCompare`. Returns `nil` when equal.
-  private func compareTitlesAscending(_ lhs: String, _ rhs: String) -> Bool? {
-    let result = lhs.localizedStandardCompare(rhs)
-    return result == .orderedSame ? nil : result == .orderedAscending
-  }
-
-  /// Deterministic tiebreaker using the lexicographic order of `providerID/nativeID`.
-  private func compareRefs(_ lhs: TaskRef, _ rhs: TaskRef) -> Bool {
-    "\(lhs.providerID)/\(lhs.nativeID)" < "\(rhs.providerID)/\(rhs.nativeID)"
   }
 
   // MARK: - Persistence
